@@ -1,6 +1,7 @@
 {
   "name": "event-emitters",
   "version": "1.0.6",
+  "type": "module",
   "description": "event emitters",
   "author": "insidewhy <github@chilon.net>",
   "license": "ISC",
@@ -35,18 +36,21 @@
     "jest-watch": "yarn jest --watchAll"
   },
   "devDependencies": {
-    "@types/jest": "26.0.22",
-    "@types/node": "14.14.37",
-    "@typescript-eslint/eslint-plugin": "4.20.0",
-    "@typescript-eslint/parser": "4.20.0",
-    "concurrently": "6.0.0",
-    "eslint": "7.23.0",
-    "eslint-config-prettier": "8.1.0",
-    "husky": "6.0.0",
-    "jest": "26.6.3",
-    "prettier": "2.2.1",
-    "pretty-quick": "3.1.0",
-    "rimraf": "3.0.2",
-    "typescript": "4.2.3"
+    "@types/jest": "29.5.12",
+    "@types/node": "20.12.12",
+    "concurrently": "8.2.2",
+    "eslint": "9.3.0",
+    "eslint-config-prettier": "9.1.0",
+    "husky": "9.0.11",
+    "jest": "29.7.0",
+    "prettier": "3.2.5",
+    "pretty-quick": "4.0.0",
+    "rimraf": "5.0.7",
+    "typescript": "5.4.5",
+    "typescript-eslint": "^7.10.0"
+  },
+  "volta": {
+    "node": "22.2.0",
+    "yarn": "1.22.22"
   }
 }
