@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   testRegex: 'spec\\.js$',
   roots: ['<rootDir>/dist.es5'],
 }
